Move Jackson support to a separate class
